drone_safety_landing_app/requirements.txt
# Python dependencies
ultralytics==8.0.203
opencv-python
Pillow
tk
tkintermapview
dronekit
pyserial

